[refactor] : atoms 컴포넌트 카테고리별로 폴더 분리
