+++ src/components/molecules/BankForm.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import { useMutation } from '@tanstack/react-query';
-import Button from '../atoms/Button';
+import Button from '../atoms/button/Button';
 import banks from '../../constant/bank';
 import routes from '../../constant/routes';
 import registerBank from '../../apis/register';
