+++ src/components/templates/PickerMatch.jsx
@@ -1,5 +1,5 @@
-import Button from '../atoms/Button';
-import OtherNav from '../atoms/OtherNav';
+import Button from '../atoms/button/Button';
+import OtherNav from '../atoms/nav/OtherNav';
 import Info from '../organisms/Info';
 import Location from '../organisms/Location';
