+++ src/pages/PostWritePage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { useForm, FormProvider } from 'react-hook-form';
 import { useMutation } from '@tanstack/react-query';
 import Swal from 'sweetalert2';
-import OtherNav from '../components/atoms/OtherNav';
+import OtherNav from '../components/atoms/nav/OtherNav';
 import BtnNavigate from '../components/molecules/BtnNavigate';
 import OrderInfo from '../components/templates/OrderInfo';
 import OrderRequest from '../components/templates/OrderRequest';
