+++ src/pages/UploadStudentCardPage.jsx
@@ -4,8 +4,8 @@ import { BsUpload } from 'react-icons/bs';
 import Swal from 'sweetalert2';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { getUserAuth } from '../apis/myPage';
-import OtherNav from '../components/atoms/OtherNav';
-import Button from '../components/atoms/Button';
+import OtherNav from '../components/atoms/nav/OtherNav';
+import Button from '../components/atoms/button/Button';
 import { useNavigate } from 'react-router-dom';
 import routes from '../constant/routes';
 import uploadCard from '../apis/uploadCard';
